Final MemSimRT
There are so many simulators available to simulate different test cases for scheduling in a realtime
operating

system, e.g. Litmus-RT, Mark3, rtsim, etc., but no simulator is available for
simulating memory management algorithm for RTOS so far. Therefore, MemSimRT has been
designed to simulate various memory allocators for both SMP as well as NUMA architecture
based RTOS. This will be useful for all existing allocators as well as the allocators which
would be proposed in future.